Fix cpuprofile and infos bug
When executing the cpupprofile and infosource commands, if the information sent by the server is missing the 'method', it will cause the client to return an exception when parsing the message information.
In the current logic, it is allowed that there is no 'method' in the message, so the judgment logic that returns when 'method' is empty is deleted.

Issue:https://gitee.com/openharmony/arkcompiler_toolchain/issues/IAD5Y9

Signed-off-by: daiweihao <daiweihao1@huawei.com>
Change-Id: I149af37ea1352cb47cac9ac2aea1ff304aaaf832