Fix WebSocketTest testcase
After disconnecting the test case, resources will remain for a period of time and the port cannot be released. If a connection is established in other use cases in the future, it will result in connection failure.
Therefore, after disconnecting, let the program sleep for 3 seconds to ensure the release of resources.

Issue:https://gitee.com/openharmony/arkcompiler_toolchain/issues/IAFVKU

Signed-off-by: daiweihao <daiweihao1@huawei.com>
Change-Id: Ic06e91b987720b4fa0db5d0779490040bb91b1c8